pyproject.toml
[project]
name = "idi-pensions-funds-scraping"
version = "0.1.0"
description = "Data pipeline for extracting pension fund investments."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bs4>=0.0.2",
    "ipykernel>=7.1.0",
    "pandas>=3.0.0",
    "pdfplumber>=0.11.9",
    "pillow>=12.1.0",
    "playwright>=1.57.0",
    "pytesseract>=0.3.13",
    "requests>=2.32.5",
]

[project.scripts]
idi-scrape = "pipeline.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pipeline"]

[dependency-groups]
dev = [
    "ruff>=0.14.14",
]

[tool.ruff]
target-version = "py312"
line-length = 80

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "D",   # pydocstyle (Google convention)
    "ANN", # flake8-annotations — require type annotations
]
ignore = [
    "E501",  # line-too-long — handled by formatter
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"